// apps/admin-web/src/api/model/auditListItemDto2.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { AuditListItemDto2Metadata } from './auditListItemDto2Metadata';
import type { AuditListItemDto2Payload } from './auditListItemDto2Payload';
import type { AuditListItemDto2ResponseData } from './auditListItemDto2ResponseData';

export interface AuditListItemDto2 {
  /**
   * 유저 ID
   * @nullable
   */
  userId?: string | null;
  /**
   * 조직 ID
   * @nullable
   */
  organizationId?: string | null;
  /** 요청 URL */
  url: string;
  /** 요청 메서드 */
  method: string;
  /** 응답 상태 코드 */
  statusCode: number;
  /** 수행 시간 (ms) */
  duration: number;
  /**
   * IP 주소
   * @nullable
   */
  ip?: string | null;
  /**
   * User-Agent
   * @nullable
   */
  userAgent?: string | null;
  /**
   * 요청한 페이지 (Referer)
   * @nullable
   */
  referer?: string | null;
  /**
   * 페이로드 (마스킹됨)
   * @nullable
   */
  payload?: AuditListItemDto2Payload;
  /**
   * 응답 데이터 (마스킹됨)
   * @nullable
   */
  responseData?: AuditListItemDto2ResponseData;
  /**
   * 에러 메시지
   * @nullable
   */
  error?: string | null;
  /** 유저 이름 */
  userName?: string;
  /** 조직 이름 */
  organizationName?: string;
  id: string;
  createdAt: string;
  createdBy?: string;
  updatedAt: string;
  updatedBy?: string;
  deletedAt?: string;
  deletedBy?: string;
  metadata?: AuditListItemDto2Metadata;
}